v 0.4
- Added support for show_options_nodes option.
- Fixed bug: self.toplevel now defined in processTree.
- Fixed bugs related to @rst-skip-node and @rst-skip-tree.
- Fixed another bug in scanOption.
- scanForOptionDocParts now treats @ @rst-option and @ @rst-options identically.

There is a problem with section levels that I don't know how to solve except with a warning:

The first node at each level must not be ignored: because rst treats the first occurance of an underlining character to be the next highest section level.  For example:

- @ignore-node
..- a
- b

rst thinks that A is at a higher level than b!